// apps/api/src/modules/tenants/dto/create-tenant.dto.d.ts
export declare class CreateTenantDto {
    nombre: string;
    nit: string;
    nrc: string;
    actividadEcon: string;
    direccion: {
        departamento: string;
        municipio: string;
        complemento: string;
    };
    telefono: string;
    correo: string;
    nombreComercial?: string;
}
//# sourceMappingURL=create-tenant.dto.d.ts.map